requirements.txt
gunicorn
Django==1.10.5
django-braces==1.11.0
django-oauth-toolkit==0.12.0
djangorestframework==3.5.4
oauthlib==2.0.1
six==1.10.0
whitenoise==3.3.0
